requirements.txt
SQLAlchemy
Flask
flask-wtf
python-dotenv
flask-login
werkzeug
itsdangerous
requests
urllib3<2.0
WTForms